gui: Split into files
